Add authentication to the blog website (#2)
* Create user Controller in backend

* Add unique tag to user model

* Rewrite signup in a static function of the model

* Add validator package

* Add checks during signup

* Install jwt

* Create token when signup

* Add login static check to the model

* Add login controller function

The function sends back the token if the login is successful

* Add auth context

* Add login and signup page with style

The action on form submit is not coded at the moment

* Add auth button to navbar

* Add useSignUp hook

* Add logout hook

* Add logout button to navbar

* Add login hook

* Add conditional navbar depending on auth context

* Rename user object in API

* Grab the auth context from localStorage

* Clean unused function

* Add requireAuth middleware to some routes

* Update article controller

Update arcticle functions for routes that need authorization

* Add bearer authorization to createArticle

* Add bearer authaurization to delete article

* Fix ids instead of username for articles' author

* Fix minor bug when deleting article

* Remove log auth context

* Add conditional routes based on AuthContext

* Update README

* Update example.env

* Read SALT_ROUND from env

* Clean code

* Add RESET type to dispatch article function

* Reset the article context when loging out

* Redirect to / when loging out

* Redirect to / after login

* Add main gitignore

* Update README.md

* Add expiration time for JWT in the config file

Move the hardcoded expiration time variable to the .env.example file
Update the readme

* Correct typo

* Change API endpoints to avoid camelCase

Rename the byAuthor endpoint to by-author to follow web name guilines

* Rename frontend route to avoid camelCase

Rename createArticle => create-article

* Correct typo in getByAuthor endpoint controller

Replace || operator by ??
Change the order of priority in the test
The author variable will be set to:
- req.params.author if it exists (the username of the author provided in
  the URL
- req.user.username if it exists (the username of the connected user)
- throw an error otherwise

* Remove duplicated test for existing username/email

* Update error message for weak password

Add password policy to the error message

* Add more precise error logs for the login

Display which of the username or password is wrong

* Add alert error messages when deleteArticle fails

* Fix author in arctiel preview during the creation

+++ backend/controllers/articleController.js
@@ -11,11 +11,9 @@ const getAll = async (req, res) => {
 };
 
 const getByAuthor = async (req, res) => {
-  let author;
-  if (!req.params.author) {
-    author = "alex";
-  } else {
-    author = req.params.author;
+  const author = req.params.author ?? req.user?.usermame
+  if (!author) {
+    res.status(404).json({error: "You must login or provide an author" });
   }
 
   try {
@@ -42,7 +40,8 @@ const getById = async (req, res) => {
 };
 
 const create = async (req, res) => {
-  const { title, author, body, summary } = req.body;
+  const { title, body, summary } = req.body;
+  const author = req.user.username;
   try {
     const article = await Article.create({ title, author, body, summary });
     res.status(200).json(article);
@@ -53,17 +52,21 @@ const create = async (req, res) => {
 
 const deleteById = async (req, res) => {
   const id = req.params.id;
+  const { username } = req.user;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({ error: "No such article" });
   }
-  const article = await Article.findByIdAndDelete(id);
+  const { author } = await Article.findById(id);
 
-  if (!article) {
+  if (!author) {
     res.status(400).json({ error: "No such article" });
+  } else if (author !== username) {
+    res.status(401).json({ error: "Access denied" });
+  } else {
+    const article = await Article.findByIdAndDelete(id);
+    res.status(200).json(article);
   }
-
-  res.status(200).json(article);
 };
 
 module.exports = { getAll, getById, getByAuthor, create, deleteById };
